chore: move `Identifier` type to crate root
Also marks `hcl::structure::Identifier` as deprecated. This symbol will
be removed in a future release.

For most users this shouldn't be an issue since the `Identifier` type
was always available as `hcl::Identifier` and none of the usage examples
referenced to it as `hcl::structure::Identifier`.
