+++ src/structure/block.rs
@@ -1,6 +1,7 @@
 //! Types to represent and build HCL blocks.
 
-use super::{Attribute, Body, BodyBuilder, Identifier, IntoNodeMap, Structure, Value};
+use super::{Attribute, Body, BodyBuilder, IntoNodeMap, Structure};
+use crate::{Identifier, Value};
 use serde::{Deserialize, Serialize};
 
 /// Represents an HCL block which consists of a block identifier, zero or more block labels and a
