+++ src/structure/for_expr.rs
@@ -1,4 +1,5 @@
-use super::{Expression, Identifier};
+use super::Expression;
+use crate::Identifier;
 use serde::{Deserialize, Serialize};
 
 /// A for expression is a construct for constructing a collection by projecting the items from
