+++ src/structure/tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::Identifier;
 use pretty_assertions::assert_eq;
 use serde_json::json;
